Trying to solve merge ui

+++ Final_Fantasy_Mystery_World/j1UIManager.cpp
@@ -75,6 +75,7 @@ bool j1UIManager::PostUpdate()
 			(*item)->DebugDraw();
 		}
 	}
+	tree.clear();
 
 	return ret;
 }
@@ -148,33 +149,34 @@ void j1UIManager::CreateScreen()
 
 bool j1UIManager::DeleteUIElement(GUI * element)
 {
-	std::list<GUI*>::iterator item_ui = std::find(ui_list.begin(), ui_list.end(), element);
-	if (item_ui != ui_list.end()) {															//if element doesn't find in ui list it cannot be deleted
+	if (element != nullptr) {
+		std::list<GUI*>::iterator item_ui = std::find(ui_list.begin(), ui_list.end(), element);
+		if (item_ui != ui_list.end()) {															//if element doesn't find in ui list it cannot be deleted
 
-		std::list<GUI*> tree;
-		BFS(tree, element);		//fills a list from element to delete to its childs using BFS algorithm
+			std::list<GUI*> tree;
+			BFS(tree, element);		//fills a list from element to delete to its childs using BFS algorithm
 
-		for (std::list<GUI*>::reverse_iterator item_tree = tree.rbegin(); item_tree != tree.rend(); ++item_tree) {	//iterate list from bottom to top
-			if (item_tree == tree.rend() && (*item_tree)->parent != nullptr) {				/*In case the item we will delete is the first element of the tree
-																					we have to delete him first from its parent child list
-																					the reason why we don't made that for other nodes is becuase
-																					other nodes and its parents will be deleted for complete*/
-				std::list<GUI*>::iterator this_on_child = std::find((*item_tree)->parent->childs.begin(), (*item_tree)->parent->childs.end(), *item_tree);
-				if (this_on_child != (*item_tree)->parent->childs.end()) {
-					(*item_tree)->parent->childs.remove(*this_on_child);
+			for (std::list<GUI*>::reverse_iterator item_tree = tree.rbegin(); item_tree != tree.rend(); ++item_tree) {	//iterate list from bottom to top
+				if (item_tree == tree.rend() && (*item_tree)->parent != nullptr) {				/*In case the item we will delete is the first element of the tree
+																						we have to delete him first from its parent child list
+																						the reason why we don't made that for other nodes is becuase
+																						other nodes and its parents will be deleted for complete*/
+					std::list<GUI*>::iterator this_on_child = std::find((*item_tree)->parent->childs.begin(), (*item_tree)->parent->childs.end(), *item_tree);
+					if (this_on_child != (*item_tree)->parent->childs.end()) {
+						(*item_tree)->parent->childs.remove(*this_on_child);
+					}
+				}
+				std::list<GUI*>::iterator elem = std::find(ui_list.begin(), ui_list.end(), *item_tree);	//find item on ui objects list
+				if (elem != ui_list.end() && *elem != nullptr) {						//if it is valid
+					delete *elem;
+					*elem = nullptr;						//delete from list
+															//delete item->data;						//and deallocate memory
 				}
 			}
-			std::list<GUI*>::iterator elem = std::find(ui_list.begin(),ui_list.end(),*item_tree);	//find item on ui objects list
-			if (elem != ui_list.end()) {						//if it is valid
-				ui_list.remove(*elem);
-				delete *elem;
-				*elem = nullptr;						//delete from list
-														//delete item->data;						//and deallocate memory
-			}
-		}
-		tree.clear();
+			tree.clear();
 
-		return true;
+			return true;
+		}
 	}
 
 	LOG("Element not found to delete");
@@ -192,6 +194,7 @@ void j1UIManager::BFS(std::list<GUI*>& visited, GUI * elem)
 		while (frontier.empty() == false) {
 			if ((item = frontier.front()) != nullptr) {			//Pop las item of array
 				frontier.pop();
+				if(item->childs.empty() == false)
 					for (std::list<GUI*>::iterator it = item->childs.begin(); it != item->childs.end(); ++it) { //iterate for all childs of node
 						if (std::find(visited.begin(),visited.end(),*it) == visited.end()) {	//if child is not on visited list we added on it and on prontier to search its childs
 							frontier.push(*it);
@@ -236,6 +239,7 @@ bool j1UIManager::GetElemOnMouse(int x, int y, GUI *& element)
 			}
 		}
 	}
+	tree.clear();
 
 	return false;
 }
